Consolidate flow physics classes with a shared base class refs #27280 #25642:wq
