+++ modules/navier_stokes/src/physics/WCNSFVFlowPhysics.C
@@ -16,38 +16,16 @@
 #include "MapConversionUtils.h"
 #include "NS.h"
 
-registerNavierStokesPhysicsBaseTasks("NavierStokesApp", WCNSFVFlowPhysics);
-registerMooseAction("NavierStokesApp", WCNSFVFlowPhysics, "add_variable");
-registerMooseAction("NavierStokesApp", WCNSFVFlowPhysics, "add_ic");
+registerWCNSFVFlowPhysicsBaseTasks("NavierStokesApp", WCNSFVFlowPhysics);
 registerMooseAction("NavierStokesApp", WCNSFVFlowPhysics, "add_fv_kernel");
 registerMooseAction("NavierStokesApp", WCNSFVFlowPhysics, "add_fv_bc");
-registerMooseAction("NavierStokesApp", WCNSFVFlowPhysics, "add_material");
-registerMooseAction("NavierStokesApp", WCNSFVFlowPhysics, "add_user_object");
-registerMooseAction("NavierStokesApp", WCNSFVFlowPhysics, "add_corrector");
-registerMooseAction("NavierStokesApp", WCNSFVFlowPhysics, "add_postprocessor");
-registerMooseAction("NavierStokesApp", WCNSFVFlowPhysics, "get_turbulence_physics");
 
 InputParameters
 WCNSFVFlowPhysics::validParams()
 {
-  InputParameters params = NavierStokesPhysicsBase::validParams();
+  InputParameters params = WCNSFVFlowPhysicsBase::validParams();
   params.addClassDescription(
       "Define the Navier Stokes weakly-compressible mass and momentum equations");
-
-  params += NSFVBase::commonMomentumEquationParams();
-  params.addParam<bool>("add_flow_equations",
-                        true,
-                        "Whether to add the flow equations. This parameter is not necessary when "
-                        "using the Physics syntax");
-
-  // We pull in parameters from various flow objects. This helps make sure the parameters are
-  // spelled the same way and match the evolution of other objects.
-  // If we remove these objects, or change their parameters, these parameters should be updated
-  // Downstream actions must either implement all these options, or redefine the parameter with
-  // a restricted MooseEnum, or place an error in the constructor for unsupported configurations
-  // We mostly pull the boundary parameters from NSFV Action
-
-  params += NSFVBase::commonNavierStokesFlowParams();
 
   // Time derivative term correction
   params.transferParam<bool>(INSFVTimeKernel::validParams(), "contribute_to_rc");
@@ -58,35 +36,8 @@ WCNSFVFlowPhysics::validParams()
       "density_for_gravity_terms",
       "If specified, replaces the 'density' for the Boussinesq and gravity momentum kernels");
 
-  // Most downstream physics implementations are valid for porous media too
-  // If yours is not, please remember to disable the 'porous_medium_treatment' parameter
-  params.transferParam<bool>(NSFVBase::validParams(), "porous_medium_treatment");
-  params.transferParam<MooseFunctorName>(NSFVBase::validParams(), "porosity");
+  // Additional porous media parameters
   params.transferParam<unsigned short>(NSFVBase::validParams(), "porosity_smoothing_layers");
-
-  // Momentum boundary conditions are important for advection problems as well
-  params += NSFVBase::commonMomentumBoundaryTypesParams();
-  params.addParam<bool>(
-      "include_deviatoric_stress",
-      false,
-      "Whether to include the full expansion (the transposed term as well) of the stress tensor");
-  params.addParam<std::vector<MooseFunctorName>>(
-      "momentum_wall_functors",
-      {},
-      "Functors for each component of the velocity value on walls. This is only necessary for the "
-      "fixed-velocity momentum wall types.");
-
-  // Specify the weakly compressible boundary flux information. They are used for specifying in flux
-  // boundary conditions for advection physics in WCNSFV
-  params += NSFVBase::commonMomentumBoundaryFluxesParams();
-
-  // New functor boundary conditions
-  params.deprecateParam("momentum_inlet_function", "momentum_inlet_functors", "01/01/2025");
-  params.deprecateParam("pressure_function", "pressure_functors", "01/01/2025");
-
-  // Initialization parameters
-  params.transferParam<std::vector<FunctionName>>(NSFVBase::validParams(), "initial_velocity");
-  params.transferParam<FunctionName>(NSFVBase::validParams(), "initial_pressure");
 
   // Techniques to limit or remove oscillations at porosity jump interfaces
   params.transferParam<MooseEnum>(NSFVBase::validParams(), "porosity_interface_pressure_treatment");
@@ -121,16 +72,6 @@ WCNSFVFlowPhysics::validParams()
   params.transferParam<Real>(NSFVBase::validParams(), "momentum_scaling");
 
   // Parameter groups
-  params.addParamNamesToGroup(
-      "velocity_variable pressure_variable initial_pressure initial_velocity", "Variables");
-  params.addParamNamesToGroup("density dynamic_viscosity characteristic_speed",
-                              "Material properties");
-  params.addParamNamesToGroup("inlet_boundaries momentum_inlet_types momentum_inlet_functors",
-                              "Inlet boundary conditions");
-  params.addParamNamesToGroup("outlet_boundaries momentum_outlet_types pressure_functors",
-                              "Outlet boundary conditions");
-  params.addParamNamesToGroup("wall_boundaries momentum_wall_types momentum_wall_functors",
-                              "Wall boundary conditions");
   params.addParamNamesToGroup("coupled_turbulence_physics", "Coupled Physics");
   params.addParamNamesToGroup(
       "porosity_interface_pressure_treatment pressure_allow_expansion_on_bernoulli_faces "
@@ -141,18 +82,12 @@ WCNSFVFlowPhysics::validParams()
       "mass_advection_interpolation momentum_advection_interpolation "
       "pressure_two_term_bc_expansion momentum_two_term_bc_expansion mass_scaling momentum_scaling",
       "Numerical scheme");
-  // TODO Add default preconditioning and move scaling parameters to a preconditioning group
-  params.addParamNamesToGroup("thermal_expansion", "Gravity treatment");
 
   return params;
 }
 
 WCNSFVFlowPhysics::WCNSFVFlowPhysics(const InputParameters & parameters)
-  : NavierStokesPhysicsBase(parameters),
-    _has_flow_equations(getParam<bool>("add_flow_equations")),
-    _compressibility(getParam<MooseEnum>("compressibility")),
-    _porous_medium_treatment(getParam<bool>("porous_medium_treatment")),
-    _porosity_name(getParam<MooseFunctorName>("porosity")),
+  : WCNSFVFlowPhysicsBase(parameters),
     _flow_porosity_functor_name(isParamValid("porosity_smoothing_layers") &&
                                         getParam<unsigned short>("porosity_smoothing_layers")
                                     ? NS::smoothed_porosity
@@ -160,43 +95,10 @@ WCNSFVFlowPhysics::WCNSFVFlowPhysics(const InputParameters & parameters)
     _porosity_smoothing_layers(isParamValid("porosity_smoothing_layers")
                                    ? getParam<unsigned short>("porosity_smoothing_layers")
                                    : 0),
-    _velocity_names(
-        isParamValid("velocity_variable")
-            ? getParam<std::vector<std::string>>("velocity_variable")
-            : (_porous_medium_treatment
-                   ? std::vector<std::string>(NS::superficial_velocity_vector,
-                                              NS::superficial_velocity_vector + 3)
-                   : std::vector<std::string>(NS::velocity_vector, NS::velocity_vector + 3))),
-    _pressure_name(isParamValid("pressure_variable")
-                       ? getParam<NonlinearVariableName>("pressure_variable")
-                       : NS::pressure),
-    _fluid_temperature_name(isParamValid("fluid_temperature_variable")
-                                ? getParam<NonlinearVariableName>("fluid_temperature_variable")
-                                : NS::T_fluid),
-    _density_name(getParam<MooseFunctorName>("density")),
-    _density_gravity_name(isParamValid("density_gravity")
-                              ? getParam<MooseFunctorName>("density_gravity")
-                              : getParam<MooseFunctorName>("density")),
-    _dynamic_viscosity_name(getParam<MooseFunctorName>("dynamic_viscosity")),
-    _velocity_interpolation(getParam<MooseEnum>("velocity_interpolation")),
-    _momentum_face_interpolation(getParam<MooseEnum>("momentum_advection_interpolation")),
-    _inlet_boundaries(getParam<std::vector<BoundaryName>>("inlet_boundaries")),
-    _outlet_boundaries(getParam<std::vector<BoundaryName>>("outlet_boundaries")),
-    _wall_boundaries(getParam<std::vector<BoundaryName>>("wall_boundaries")),
-    _momentum_wall_types(getParam<MultiMooseEnum>("momentum_wall_types")),
-    _momentum_wall_functors(getParam<std::vector<MooseFunctorName>>("momentum_wall_functors")),
-    _flux_inlet_pps(getParam<std::vector<PostprocessorName>>("flux_inlet_pps")),
-    _flux_inlet_directions(getParam<std::vector<Point>>("flux_inlet_directions")),
     _friction_blocks(getParam<std::vector<std::vector<SubdomainName>>>("friction_blocks")),
     _friction_types(getParam<std::vector<std::vector<std::string>>>("friction_types")),
     _friction_coeffs(getParam<std::vector<std::vector<std::string>>>("friction_coeffs"))
 {
-  // Inlet boundary parameter checking
-  checkSecondParamSetOnlyIfFirstOneSet("flux_inlet_pps", "flux_inlet_directions");
-  if (_flux_inlet_directions.size())
-    checkVectorParamsSameLengthIfSet<PostprocessorName, Point>("flux_inlet_pps",
-                                                               "flux_inlet_directions");
-
   // Most likely to be a mistake
   if (getParam<bool>("pin_pressure") &&
       getParam<std::vector<MooseFunctorName>>("pressure_functors").size())
@@ -227,10 +129,6 @@ WCNSFVFlowPhysics::WCNSFVFlowPhysics(const InputParameters & parameters)
                      std::string(getParam<MooseEnum>("pinned_pressure_type")) + ".");
   }
 
-  // Boussinesq parameters checks
-  checkSecondParamSetOnlyIfFirstOneTrue("boussinesq_approximation", "ref_temperature");
-  checkSecondParamSetOnlyIfFirstOneTrue("boussinesq_approximation", "thermal_expansion");
-
   // Porosity correction checks
   checkSecondParamSetOnlyIfFirstOneTrue("porous_medium_treatment", "use_friction_correction");
   checkSecondParamSetOnlyIfFirstOneTrue("use_friction_correction", "consistent_scaling");
@@ -241,83 +139,8 @@ WCNSFVFlowPhysics::WCNSFVFlowPhysics(const InputParameters & parameters)
                             "bernoulli",
                             {"pressure_allow_expansion_on_bernoulli_faces"});
 
-  // Boundary parameters checking
-  checkVectorParamAndMultiMooseEnumLength<BoundaryName>("inlet_boundaries", "momentum_inlet_types");
-  checkVectorParamAndMultiMooseEnumLength<BoundaryName>("outlet_boundaries",
-                                                        "momentum_outlet_types");
-  checkVectorParamAndMultiMooseEnumLength<BoundaryName>("wall_boundaries", "momentum_wall_types");
-  checkVectorParamLengthSameAsCombinedOthers<BoundaryName,
-                                             std::vector<MooseFunctorName>,
-                                             PostprocessorName>(
-      "inlet_boundaries", "momentum_inlet_functors", "flux_inlet_pps");
-  checkVectorParamsNoOverlap<BoundaryName>(
-      {"inlet_boundaries", "outlet_boundaries", "wall_boundaries"});
-
   // Porous media parameters
-  checkSecondParamSetOnlyIfFirstOneTrue("porous_medium_treatment", "porosity");
   checkSecondParamSetOnlyIfFirstOneTrue("porous_medium_treatment", "porosity_smoothing_layers");
-
-  if (_define_variables && _porous_medium_treatment)
-    for (const auto & name : NS::velocity_vector)
-    {
-      const auto & it = std::find(_velocity_names.begin(), _velocity_names.end(), name);
-      if (it != _velocity_names.end())
-        paramError("velocity_variable",
-                   "For porous medium simulations, functor name " + *it +
-                       " is already reserved for the automatically-computed interstitial velocity. "
-                       "Please choose another name for your external velocity variable!");
-    }
-
-  // Create maps for boundary-restricted parameters
-  _momentum_inlet_types = Moose::createMapFromVectorAndMultiMooseEnum<BoundaryName>(
-      _inlet_boundaries, getParam<MultiMooseEnum>("momentum_inlet_types"));
-  _momentum_outlet_types = Moose::createMapFromVectorAndMultiMooseEnum<BoundaryName>(
-      _outlet_boundaries, getParam<MultiMooseEnum>("momentum_outlet_types"));
-  if (isParamSetByUser("momentum_inlet_functors"))
-  {
-    // Not all inlet boundary types require the specification of an inlet functor
-    std::vector<BoundaryName> inlet_boundaries_with_functors;
-    for (const auto & boundary : _inlet_boundaries)
-      if (libmesh_map_find(_momentum_inlet_types, boundary) == "fixed-velocity" ||
-          libmesh_map_find(_momentum_inlet_types, boundary) == "fixed-pressure")
-        inlet_boundaries_with_functors.push_back(boundary);
-    _momentum_inlet_functors =
-        Moose::createMapFromVectors<BoundaryName, std::vector<MooseFunctorName>>(
-            inlet_boundaries_with_functors,
-            getParam<std::vector<std::vector<MooseFunctorName>>>("momentum_inlet_functors"));
-  }
-  if (isParamSetByUser("pressure_functors"))
-  {
-    // Not all inlet boundary types require the specification of an inlet functor
-    std::vector<BoundaryName> outlet_boundaries_with_functors;
-    for (const auto & boundary : _outlet_boundaries)
-      if (libmesh_map_find(_momentum_outlet_types, boundary) == "fixed-pressure-zero-gradient" ||
-          libmesh_map_find(_momentum_outlet_types, boundary) == "fixed-pressure")
-        outlet_boundaries_with_functors.push_back(boundary);
-    const auto & pressure_functors = getParam<std::vector<MooseFunctorName>>("pressure_functors");
-    if (outlet_boundaries_with_functors.size() != pressure_functors.size())
-      paramError("pressure_functors",
-                 "Size (" + std::to_string(pressure_functors.size()) +
-                     ") is not the same as the number of pressure outlet boundaries in "
-                     "'fixed-pressure/fixed-pressure-zero-gradient' (size " +
-                     std::to_string(outlet_boundaries_with_functors.size()) + ")");
-    _pressure_functors = Moose::createMapFromVectors<BoundaryName, MooseFunctorName>(
-        outlet_boundaries_with_functors, pressure_functors);
-  }
-}
-
-void
-WCNSFVFlowPhysics::initializePhysicsAdditional()
-{
-  getProblem().needFV();
-}
-
-void
-WCNSFVFlowPhysics::actOnAdditionalTasks()
-{
-  // Turbulence physics would not be initialized before this task
-  if (_current_task == "get_turbulence_physics")
-    _turbulence_physics = getCoupledTurbulencePhysics();
 }
 
 void
@@ -852,14 +675,6 @@ WCNSFVFlowPhysics::addINSMomentumFrictionKernels()
 }
 
 void
-WCNSFVFlowPhysics::addFVBCs()
-{
-  addINSInletBC();
-  addINSOutletBC();
-  addINSWallsBC();
-}
-
-void
 WCNSFVFlowPhysics::addINSInletBC()
 {
   // Check the size of the BC parameters
@@ -1233,15 +1048,6 @@ WCNSFVFlowPhysics::addCorrectors()
   }
 }
 
-void
-WCNSFVFlowPhysics::addMaterials()
-{
-  if (hasForchheimerFriction() || _porous_medium_treatment)
-    addPorousMediumSpeedMaterial();
-  else
-    addNonPorousMediumSpeedMaterial();
-}
-
 bool
 WCNSFVFlowPhysics::hasForchheimerFriction() const
 {
@@ -1252,102 +1058,10 @@ WCNSFVFlowPhysics::hasForchheimerFriction() const
   return false;
 }
 
-void
-WCNSFVFlowPhysics::addPorousMediumSpeedMaterial()
-{
-  InputParameters params = getFactory().getValidParams("PINSFVSpeedFunctorMaterial");
-  assignBlocks(params, _blocks);
-
-  for (unsigned int dim_i = 0; dim_i < dimension(); ++dim_i)
-    params.set<MooseFunctorName>(NS::superficial_velocity_vector[dim_i]) = _velocity_names[dim_i];
-  if (_porous_medium_treatment)
-    params.set<MooseFunctorName>(NS::porosity) = _flow_porosity_functor_name;
-  else
-    params.set<MooseFunctorName>(NS::porosity) = "1";
-  params.set<bool>("define_interstitial_velocity_components") = _porous_medium_treatment;
-
-  getProblem().addMaterial("PINSFVSpeedFunctorMaterial", prefix() + "pins_speed_material", params);
-}
-
-void
-WCNSFVFlowPhysics::addNonPorousMediumSpeedMaterial()
-{
-  const std::string class_name = "ADVectorMagnitudeFunctorMaterial";
-  InputParameters params = getFactory().getValidParams(class_name);
-  assignBlocks(params, _blocks);
-
-  const std::vector<std::string> param_names{"x_functor", "y_functor", "z_functor"};
-  for (unsigned int dim_i = 0; dim_i < dimension(); ++dim_i)
-    params.set<MooseFunctorName>(param_names[dim_i]) = _velocity_names[dim_i];
-  params.set<MooseFunctorName>("vector_magnitude_name") = NS::speed;
-
-  getProblem().addMaterial(class_name, prefix() + "ins_speed_material", params);
-}
-
-void
-WCNSFVFlowPhysics::addInitialConditions()
-{
-  if (!_define_variables && parameters().isParamSetByUser("initial_velocity") &&
-      parameters().isParamSetByUser("velocity_variable") &&
-      getParam<std::vector<FunctionName>>("initial_velocity").size() != 0)
-    // TODO: Rework and remove this last statement once the NSFV action is removed
-    paramError("initial_velocity",
-               "Velocity is defined externally of WCNSFVFlowPhysics, so should the inital "
-               "conditions");
-  if (!_define_variables && parameters().isParamSetByUser("initial_pressure") &&
-      parameters().isParamSetByUser("pressure_variable"))
-    paramError("initial_pressure",
-               "Pressure is defined externally of WCNSFVFlowPhysics, so should the inital "
-               "condition");
-
-  // Check dimension
-  if (getParam<std::vector<FunctionName>>("initial_velocity").size() != dimension() &&
-      getParam<std::vector<FunctionName>>("initial_velocity").size() != 3 &&
-      getParam<std::vector<FunctionName>>("initial_velocity").size() != 0)
-    // TODO: Rework and remove this last statement once the NSFV action is removed
-    paramError(
-        "initial_velocity",
-        "The number of velocity components in the WCNSFVFlowPhysics initial condition is not " +
-            std::to_string(dimension()) + " or 3!");
-
-  // do not set initial conditions if we load from file
-  if (getParam<bool>("initialize_variables_from_mesh_file"))
-    return;
-  // do not set initial conditions if we are not defining variables
-  if (!_define_variables)
-    return;
-
-  InputParameters params = getFactory().getValidParams("FunctionIC");
-  assignBlocks(params, _blocks);
-  auto vvalue = getParam<std::vector<FunctionName>>("initial_velocity");
-
-  if (!_app.isRestarting() || parameters().isParamSetByUser("initial_velocity"))
-    for (const auto d : make_range(dimension()))
-    {
-      params.set<VariableName>("variable") = _velocity_names[d];
-      params.set<FunctionName>("function") = vvalue[d];
-
-      getProblem().addInitialCondition("FunctionIC", prefix() + _velocity_names[d] + "_ic", params);
-    }
-
-  if (!_app.isRestarting() || parameters().isParamSetByUser("initial_pressure"))
-  {
-    params.set<VariableName>("variable") = _pressure_name;
-    params.set<FunctionName>("function") = getParam<FunctionName>("initial_pressure");
-
-    getProblem().addInitialCondition("FunctionIC", prefix() + _pressure_name + "_ic", params);
-  }
-}
-
 unsigned short
 WCNSFVFlowPhysics::getNumberAlgebraicGhostingLayersNeeded() const
 {
-  unsigned short ghost_layers = 2;
-  if ((_porous_medium_treatment &&
-       getParam<MooseEnum>("porosity_interface_pressure_treatment") != "automatic") ||
-      getParam<MooseEnum>("momentum_face_interpolation") == "skewness-corrected" ||
-      getParam<MooseEnum>("pressure_face_interpolation") == "skewness-corrected")
-    ghost_layers = std::max(ghost_layers, (unsigned short)3);
+  auto ghost_layers = WCNSFVFlowPhysicsBase::getNumberAlgebraicGhostingLayersNeeded();
   if (_porous_medium_treatment && isParamValid("porosity_smoothing_layers"))
     ghost_layers = std::max(getParam<unsigned short>("porosity_smoothing_layers"), ghost_layers);
   return ghost_layers;
@@ -1432,58 +1146,10 @@ WCNSFVFlowPhysics::checkRhieChowFunctorsDefined() const
     mooseError("Rhie Chow coefficient az must be provided for advection by auxiliary velocities");
 }
 
-void
-WCNSFVFlowPhysics::addPostprocessors()
-{
-  const auto momentum_inlet_types = getParam<MultiMooseEnum>("momentum_inlet_types");
-
-  for (unsigned int bc_ind = 0; bc_ind < momentum_inlet_types.size(); ++bc_ind)
-    if (momentum_inlet_types[bc_ind] == "flux-mass" ||
-        momentum_inlet_types[bc_ind] == "flux-velocity")
-    {
-      const std::string pp_type = "AreaPostprocessor";
-      InputParameters params = getFactory().getValidParams(pp_type);
-      params.set<std::vector<BoundaryName>>("boundary") = {_inlet_boundaries[bc_ind]};
-      params.set<ExecFlagEnum>("execute_on") = EXEC_INITIAL;
-
-      const auto name_pp = "area_pp_" + _inlet_boundaries[bc_ind];
-      if (!getProblem().hasUserObject(name_pp))
-        getProblem().addPostprocessor(pp_type, name_pp, params);
-    }
-}
-
 UserObjectName
 WCNSFVFlowPhysics::rhieChowUOName() const
 {
   return (_porous_medium_treatment ? +"pins_rhie_chow_interpolator" : "ins_rhie_chow_interpolator");
-}
-
-VariableName
-WCNSFVFlowPhysics::getFlowVariableName(const std::string & short_name) const
-{
-  if (short_name == NS::pressure)
-    return getPressureName();
-  else if (short_name == NS::velocity_x && dimension() > 0)
-    return getVelocityNames()[0];
-  else if (short_name == NS::velocity_y && dimension() > 1)
-    return getVelocityNames()[1];
-  else if (short_name == NS::velocity_z && dimension() > 2)
-    return getVelocityNames()[2];
-  else if (short_name == NS::temperature)
-    return getFluidTemperatureName();
-  else
-    mooseError("Short Variable name '", short_name, "' not recognized.");
-}
-
-MooseFunctorName
-WCNSFVFlowPhysics::getPorosityFunctorName(bool smoothed) const
-{
-  mooseAssert(!smoothed || !_porosity_smoothing_layers,
-              "Smooth porosity cannot be used without the smoothing treatment turned on");
-  if (smoothed)
-    return _flow_porosity_functor_name;
-  else
-    return _porosity_name;
 }
 
 MooseFunctorName
@@ -1529,24 +1195,4 @@ WCNSFVFlowPhysics::getLinearFrictionCoefName() const
       return "";
   }
   mooseError("Should not get here");
-}
-
-const WCNSFVTurbulencePhysics *
-WCNSFVFlowPhysics::getCoupledTurbulencePhysics() const
-{
-  // User passed it, just use that
-  if (isParamValid("coupled_turbulence_physics"))
-    return getCoupledPhysics<WCNSFVTurbulencePhysics>(
-        getParam<PhysicsName>("coupled_turbulence_physics"));
-  // Look for any physics of the right type, and check the block restriction
-  else
-  {
-    const auto all_turbulence_physics = getCoupledPhysics<const WCNSFVTurbulencePhysics>(true);
-    for (const auto physics : all_turbulence_physics)
-      if (checkBlockRestrictionIdentical(
-              physics->name(), physics->blocks(), /*error_if_not_identical=*/false))
-        return physics;
-  }
-  // Did not find one
-  return nullptr;
 }
